backend/requirements.txt
django
mysqlclient
djangorestframework
django-cors-headers
pillow
